all the pages working

+++ GamesTracker/src/main/java/com/edoatley/repository/GamesRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface GamesRepository extends CrudRepository<Game, Long> {
+public interface GamesRepository extends CrudRepository<Game, Integer> {
 
 }
